# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SecureTrack ERP API",
    "version": "1.0.0",
    "description": "Auto-generated inventory of Next.js route handlers. Auth: session cookie (Supabase) unless public/worker."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase auth session cookies"
      },
      "WorkerSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-job-secret",
        "description": "JOB_WORKER_SECRET or CRON_SECRET"
      }
    }
  },
  "paths": {
...
    "/api/attendance/devices/zkteco/iclock": {
      "get": {
        "summary": "GET /api/attendance/devices/zkteco/iclock",
        "tags": [
          "attendance",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "summary": "POST /api/attendance/devices/zkteco/iclock",
        "tags": [
          "attendance",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/audit/events": {
      "get": {
        "summary": "GET /api/audit/events",
        "tags": [
          "audit",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "summary": "POST /api/audit/events",
        "tags": [
          "audit",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/audit/reports": {
      "get": {
        "summary": "GET /api/audit/reports",
        "tags": [
          "audit",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
    "/api/audit/siem": {
      "get": {
        "summary": "GET /api/audit/siem",
        "tags": [
          "audit",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "post": {
        "summary": "POST /api/audit/siem",
        "tags": [
          "audit",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/api/email/send": {
      "post": {
        "summary": "POST /api/email/send",
        "tags": [
          "email",
          "authenticated"
        ],
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
  }
}